pyproject.toml
[project]
name = "comet-api"
version = "1.0.0"
description = "Python Fast API for the Comet Starter App"
readme = "README.md"
license = {file = "LICENSE.md"}
dependencies = [
  "alembic",
  "fastapi",
  "httpx",
  "psycopg2-binary",
  "pydantic-settings",
  "python-dotenv",
  "python-jose[cryptography]",
  "requests",
  "sqlalchemy",
  "uvicorn[standard]",
]

[project.optional-dependencies]
dev = ["coverage", "pytest", "pytest-asyncio", "ruff"]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
select = ["B", "C4", "EXE", "F", "E", "ISC", "ICN", "INP", "PIE", "SIM", "W", "T20", "UP", "T10", "G", "C90", "ERA"]
ignore = ["B008", "SIM102"]
fixable = ["F", "E", "B", "C4", "EXE", "ISC", "ICN", "INP", "PIE", "SIM", "W", "T20", "UP"]
unfixable = []

exclude = [
    ".git",
    ".mypy_cache",
    ".pre-commit-cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    "venv",
    "docs",
    "__pycache",
    "**/migrations/*",
]

line-length = 88
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
